+++ TTHAnalysis/python/tools/leptonJetReCleaner.py
@@ -1,8 +1,6 @@
 from CMGTools.TTHAnalysis.treeReAnalyzer import *
 from PhysicsTools.HeppyCore.utils.deltar import matchObjectCollection3
 import ROOT
-import numpy as num
-from math import copysign
 
 class MyVarProxy:
     def __init__(self,lep):
@@ -157,34 +155,6 @@ class LeptonJetReCleaner:
         ret["mhtJet"+self.strJetPt+"j"+postfix] = mhtJetPtvec.Pt()
         return cleanjets
 
-    def bestZ1TL(self,lepsl,lepst,cut=lambda lep:True):
-          pairs = []
-          for l1 in lepst:
-            if not cut(l1): continue
-            for l2 in lepsl:
-                if not cut(l2): continue
-                if l1.pdgId == -l2.pdgId:
-                   mz = (l1.p4() + l2.p4()).M()
-                   diff = abs(mz-91.2)
-                   pairs.append( (diff,mz) )
-          if len(pairs):
-              pairs.sort()
-              return pairs[0][1]
-          return 0.
-    def minMllTL(self, lepsl, lepst, bothcut=lambda lep:True, onecut=lambda lep:True, paircut=lambda lep1,lep2:True):
-            pairs = []
-            for l1 in lepst:
-                if not bothcut(l1): continue
-                for l2 in lepsl:
-                    if l2 == l1 or not bothcut(l2): continue
-                    if not onecut(l1) and not onecut(l2): continue
-                    if not paircut(l1,l2): continue
-                    mll = (l1.p4() + l2.p4()).M()
-                    pairs.append(mll)
-            if len(pairs):
-                return min(pairs)
-            return -1
-
     def __call__(self,event):
         self.ev = event
         fullret = {}
@@ -241,6 +211,35 @@ class LeptonJetReCleaner:
         return fullret
 
 
+def bestZ1TL(self,lepsl,lepst,cut=lambda lep:True):
+      pairs = []
+      for l1 in lepst:
+        if not cut(l1): continue
+        for l2 in lepsl:
+            if not cut(l2): continue
+            if l1.pdgId == -l2.pdgId:
+               mz = (l1.p4() + l2.p4()).M()
+               diff = abs(mz-91.2)
+               pairs.append( (diff,mz) )
+      if len(pairs):
+          pairs.sort()
+          return pairs[0][1]
+      return 0.
+
+def minMllTL(self, lepsl, lepst, bothcut=lambda lep:True, onecut=lambda lep:True, paircut=lambda lep1,lep2:True):
+        pairs = []
+        for l1 in lepst:
+            if not bothcut(l1): continue
+            for l2 in lepsl:
+                if l2 == l1 or not bothcut(l2): continue
+                if not onecut(l1) and not onecut(l2): continue
+                if not paircut(l1,l2): continue
+                mll = (l1.p4() + l2.p4()).M()
+                pairs.append(mll)
+        if len(pairs):
+            return min(pairs)
+        return -1
+
 def passMllVeto(l1, l2, mZmin, mZmax, isOSSF ):
     if  l1.pdgId == -l2.pdgId or not isOSSF:
         mz = (l1.p4() + l2.p4()).M()
@@ -262,167 +261,6 @@ def passTripleMllVeto(l1, l2, l3, mZmin, mZmax, isOSSF ):
     if all(ls): return True
     return False
 
-def _tthlep_lepId(lep):
-        #if lep.pt <= 10: return False
-        if abs(lep.pdgId) == 13:
-            if lep.pt <= 5: return False
-            if not lep.tightCharge >= 1: return False
-            return True
-        elif abs(lep.pdgId) == 11:
-            if lep.pt <= 7: return False
-            if not (lep.convVeto and lep.lostHits == 0 and lep.tightCharge >= 2): 
-                return False
-            return True
-        return False
-
-def _ttH_idEmu_cuts_E2(lep):
-    if (abs(lep.pdgId)!=11): return True
-    if (lep.hadronicOverEm>=(0.10-0.03*(abs(lep.etaSc)>1.479))): return False
-    if (abs(lep.dEtaScTrkIn)>=(0.01-0.002*(abs(lep.etaSc)>1.479))): return False
-    if (abs(lep.dPhiScTrkIn)>=(0.04+0.03*(abs(lep.etaSc)>1.479))): return False
-    if (lep.eInvMinusPInv<=-0.05): return False
-    if (lep.eInvMinusPInv>=(0.01-0.005*(abs(lep.etaSc)>1.479))): return False
-    if (lep.sigmaIEtaIEta>=(0.011+0.019*(abs(lep.etaSc)>1.479))): return False
-    return True
-
-def _susy2lss_lepId_CBloose(lep):
-        if abs(lep.pdgId) == 13:
-            if lep.pt <= 5: return False
-            return True #lep.mediumMuonId > 0
-        elif abs(lep.pdgId) == 11:
-            if lep.pt <= 7: return False
-            if not (lep.convVeto and lep.lostHits <= 1): 
-                return False
-            if not lep.mvaIdSpring15 > -0.70+(-0.83+0.70)*(abs(lep.etaSc)>0.8)+(-0.92+0.83)*(abs(lep.etaSc)>1.479):
-                return False
-            if not _susy2lss_idEmu_cuts(lep): return False
-            return True
-        return False
-
-def _susy2lss_lepConePt1015(lep):
-    if lep.conept <= (10 if abs(lep.pdgId)==13 else 15): return False
-    return True
-
-def _susy2lss_lepId_loosestFO(lep):
-    if not _susy2lss_lepId_CBloose(lep): return False
-    if abs(lep.pdgId) == 13:
-        return lep.mediumMuonId > 0 and lep.tightCharge > 0
-    elif abs(lep.pdgId) == 11:
-        return (lep.convVeto and lep.tightCharge > 1 and lep.lostHits == 0)
-    return False
-
-def _susy2lss_lepId_tighterFO(lep):
-    if not _susy2lss_lepId_loosestFO(lep): return False
-    if abs(lep.pdgId)==11:
-        if not lep.mvaIdSpring15 > -0.155+(-0.56+0.155)*(abs(lep.etaSc)>0.8)+(-0.76+0.56)*(abs(lep.etaSc)>1.479):
-            return False
-        if not _susy2lss_idIsoEmu_cuts(lep): return False
-    return True
-
-def _susy2lss_lepId_inSituLoosestFO(lep):
-    if not _susy2lss_lepId_loosestFO(lep): return False
-    if abs(lep.pdgId)==11:
-        if not lep.mvaIdSpring15 > -0.363+(-0.579+0.363)*(abs(lep.etaSc)>0.8)+(-0.623+0.579)*(abs(lep.etaSc)>1.479):
-            return False
-    return True
-
-def _susy2lss_lepId_inSituTighterFO(lep):
-    if not _susy2lss_lepId_loosestFO(lep): return False
-    if abs(lep.pdgId)==11:
-        if not lep.mvaIdSpring15 > 0.051+(-0.261-0.051)*(abs(lep.etaSc)>0.8)+(-0.403+0.261)*(abs(lep.etaSc)>1.479):
-            return False
-        if not _susy2lss_idIsoEmu_cuts(lep): return False
-    return True
-
-def _susy2lss_lepId_IPcuts(lep):
-    if not lep.sip3d<4: return False
-    if not (abs(lep.dxy)<0.05): return False
-    if not (abs(lep.dz)<0.1): return False
-    return True
-
-def _susy2lss_lepId_CB(lep):
-    if not _susy2lss_lepId_CBloose(lep): return False
-    if not _susy2lss_lepId_IPcuts(lep): return False
-    if abs(lep.pdgId) == 13:
-        return lep.mediumMuonId > 0 and lep.tightCharge > 0
-    elif abs(lep.pdgId) == 11:
-        if not (lep.convVeto and lep.tightCharge > 1 and lep.lostHits == 0): 
-            return False
-        return lep.mvaIdSpring15 > 0.87+(0.60-0.87)*(abs(lep.etaSc)>0.8)+(0.17-0.60)*(abs(lep.etaSc)>1.479)
-    return False
-
-def _susy2lss_idEmu_cuts(lep):
-    if (abs(lep.pdgId)!=11): return True
-    if (lep.sigmaIEtaIEta>=(0.011 if abs(lep.etaSc)<1.479 else 0.031)): return False
-    if (lep.hadronicOverEm>=0.08): return False
-    if (abs(lep.dEtaScTrkIn)>=0.01): return False
-    if (abs(lep.dPhiScTrkIn)>=(0.04 if abs(lep.etaSc)<1.479 else 0.08)): return False
-    if (abs(lep.eInvMinusPInv)>=0.01): return False
-    return True
-
-def _susy2lss_idIsoEmu_cuts(lep):
-    if (abs(lep.pdgId)!=11): return True
-    if not _susy2lss_idEmu_cuts(lep): return False
-    if (lep.ecalPFClusterIso>=0.45*lep.pt): return False
-    if (lep.hcalPFClusterIso>=0.25*lep.pt): return False
-    if (lep.dr03TkSumPt>=0.2*lep.pt): return False
-    return True
-
-def _susy2lss_multiIso(lep):
-        if abs(lep.pdgId) == 13: A,B,C = (0.16,0.76,7.2)
-        else:                    A,B,C = (0.12,0.80,7.2)
-        return lep.miniRelIso < A and (lep.jetPtRatiov2 > B or lep.jetPtRelv2 > C)
-
-def _susy2lss_multiIso_relaxedForInSituApp(lep):
-        if abs(lep.pdgId) == 13: A,B,C = (0.4,0.76,7.2)
-        else:                    A,B,C = (0.4,0.80,7.2)
-        return lep.miniRelIso < A and (1/lep.jetPtRatiov2 < (1/B + lep.miniRelIso) or lep.jetPtRelv2 > C)
-
-#def _susy2lss_multiIso_withMiniIsoRelaxed_ConePtJetPtRatiov2(lep):
-#        if abs(lep.pdgId) == 13: A,B,C = (0.4,0.76,7.2)
-#        else:                    A,B,C = (0.4,0.80,7.2)
-#        return lep.miniRelIso < A and (conept(lep.pt,lep.miniRelIso,lep.jetPtRatiov2,lep.jetPtRelv2,lep.pdgId,2)/lep.pt*lep.jetPtRatiov2 > B or lep.jetPtRelv2 > C)
-#def _susy2lss_multiIso_withMiniIsoRelaxed_CutForFO4(lep):
-#        if abs(lep.pdgId) == 13: A,B,C = (0.4,0.76,7.2)
-#        else:                    A,B,C = (0.4,0.80,7.2)
-#        return lep.miniRelIso < A and (1/lep.jetPtRatiov2 < (1/B + lep.miniRelIso))
-
-#def _susy2lss_lepId_CBOld(lep):
-#        if lep.pt <= 10: return False
-#        if abs(lep.pdgId) == 13:
-#            return lep.tightId > 0
-#        elif abs(lep.pdgId) == 11:
-#            return lep.tightId >= 2 and lep.convVeto and lep.tightCharge > 1 and lep.lostHits == 0
-#        return False
-
-## CH: RA7 selection for Moriond
-
-def _susy3l_multiIso(lep):
-    # CH: looser WP than for RA5 (electrons -> medium, muons -> loose)
-    if abs(lep.pdgId) == 13: A,B,C = (0.20,0.69,6.0)
-    else:                    A,B,C = (0.16,0.76,7.2)
-    return lep.miniRelIso < A and (lep.jetPtRatiov2 > B or lep.jetPtRelv2 > C)
-
-def _susy3l_lepId_loosestFO(lep):
-    # CH: the same as the 2lss one but without tightCharge
-    if not _susy2lss_lepId_CBloose(lep): return False
-    if abs(lep.pdgId) == 13:
-        return lep.mediumMuonId > 0
-    elif abs(lep.pdgId) == 11:
-        return (lep.convVeto and lep.lostHits == 0)
-    return False
-
-def _susy3l_lepId_CB(lep):
-    # CH: the same as the 2lss one but without tightCharge
-    if not _susy2lss_lepId_CBloose(lep): return False
-    if not _susy2lss_lepId_IPcuts(lep): return False
-    if abs(lep.pdgId) == 13:
-        return lep.mediumMuonId > 0
-    elif abs(lep.pdgId) == 11:
-        if not (lep.convVeto and lep.lostHits == 0): 
-            return False
-        return lep.mvaIdSpring15 > 0.87+(0.60-0.87)*(abs(lep.eta)>0.8)+(0.17-0.60)*(abs(lep.eta)>1.479)
-    return False
 
 if __name__ == '__main__':
     from sys import argv
